backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
faster-whisper
sounddevice
numpy<2.0
scipy
noisereduce
python-multipart
python-dotenv
pydub
librosa
httpx
celery
redis
pydantic-settings
pyannote.audio==3.1.1
huggingface_hub<0.24.0
torch==2.0.1
torchaudio==2.0.2
flatten-dict


structlog
asgi-correlation-id
jiwer
